Distinguish blocking and non-blocking parquet table.
