+++ src/query/storages/parquet/Cargo.toml
@@ -29,8 +29,8 @@ storages-common-table-meta = { path = "../common/table-meta" }
 async-trait = { version = "0.1.57", package = "async-trait-fn" }
 chrono = { workspace = true }
 futures = "0.3.24"
+glob = "0.3.1"
 opendal = { workspace = true }
-regex = "1.6.0"
 serde = { workspace = true }
 typetag = "0.2.3"
